addressing gh comments
